pyproject.toml
[tool.poetry]
name = "Consulting Reporting"
version = "3.1.0"
description = "HIFIS Consulting Reporting"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    {include = "cr_analysis", from = "src" },
    {include = "cr_cli", from = "src" }
]

[tool.poetry.scripts]
reporting="src.cr_cli.__main__:app"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
jsonschema = "^4.21.1"
pyyaml = "^6.0.1"
rich = "^13.7.1"
yamllint = "^1.35.1"
typer = "^0.12.3"

[tool.poetry.group.dev.dependencies]
reuse = "^3.0.2"
ipython = "^8.23.0"
pandas-stubs = "^2.2.1.240316"
types-jsonschema = "^4.21.0.20240331"
types-pyyaml = "^6.0.12.20240311"

[tool.poetry.group.gui.dependencies]
streamlit = "^1.33.0"
plotly = "^5.20.0"
streamlit-tags = "^1.2.8"

[tool.mypy-yamllint]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

